src/requirements.txt
psutil==5.7.3
mariadb==1.0.5
pycryptodome==3.9.9
pyminifier==2.1
pyarmor==6.5.5
